+++ CHANGELOG.md
@@ -1,9 +1,11 @@
 # Change log
 
 v1.2.1
+
 * Add Fork me on Github button
 * Add Google Nexus 7 (2013)
 * #13: Fix the width of device sub-menu
+* #15: Set the Ajax loading indicator at the center of screen
 
 v1.2.0 (2013/09/27)